Extract parser, storage and task list class
