+++ src/main/java/DukeException.java
@@ -1,0 +1,5 @@
+public class DukeException extends Exception {
+    DukeException(String message) {
+        super(message);
+    }
+}
